Support Symfony 7.0 (and 6.0)
The differences between Symfony 6 and 7 do not affect this codebase (yet). This means we might as well support both versions. This also adds some GHA/Composer changes to run inspections on supported PHP versions, but also on both Symfony versions. Symfony is way faster with dropping support for languages, so I added `--ignore-platform-reqs` to deal with this.